[LayoutNG] Only apply block-end margins where the block actually ends.
If we reached the end of a block in an earlier fragmentainer, but
there's overflowing content, we'll revisit this block in subsequent
fragmentainers as long as child content needs it. Repeating the
block-end margins in these fragmentainers would be wrong, though.

Use AdjustForFragmentation() only for margins, and rename it to
AdjustMarginsForFragmentation().
The fieldset layout algorithm was using this one for borders/padding,
while block and multicol layout called
AdjustBorderScrollbarPaddingForFragmentation() in container_builder_.
Become more uniform by calling that method for all in
SetupFragmentBuilderForFragmentation() (via the base layout algorithm
constructor). Since this depends on the initial fragment geometry being
set, also do that in the base layout algorithm constructor.

Bug: 829028
Change-Id: I6808d023e4e79ab489bdb3a48e3c6c47cf9db6c6